+++ Sources/ArcGISToolkit/Components/AR/ARSwiftUIView.swift
@@ -133,50 +133,3 @@ class ARSwiftUIViewProxy: NSObject, ARSessionProviding {
         )
     }
 }
-
-/// A SwiftUI version of an ARCoachingOverlayView view.
-struct ARCoachingOverlay: UIViewRepresentable {
-    /// The data source for an AR sesison.
-    var sessionProvider: ARSessionProviding?
-    /// The goal for the coaching overlay.
-    var goal: ARCoachingOverlayView.Goal
-    /// A Boolean value that indicates if coaching is in progress.
-    var active: Bool = false
-    
-    /// Controls whether the coaching is in progress.
-    /// - Parameter active: A Boolean value indicating if coaching is in progress.
-    /// - Returns: The `ARCoachingOverlay`.
-    func active(_ active: Bool) -> Self {
-        var view = self
-        view.active = active
-        return view
-    }
-    
-    /// Sets the AR session data source for the coaching overlay.
-    /// - Parameter sessionProvider: The AR session data source.
-    /// - Returns: The `ARCoachingOverlay`.
-    func sessionProvider(_ sessionProvider: ARSessionProviding) -> Self {
-        var view = self
-        view.sessionProvider = sessionProvider
-        return view
-    }
-    
-    func makeUIView(context: Context) -> ARCoachingOverlayView {
-        let view = ARCoachingOverlayView()
-        view.delegate = context.coordinator
-        view.activatesAutomatically = false
-        return view
-    }
-    
-    func updateUIView(_ uiView: ARCoachingOverlayView, context: Context) {
-        uiView.sessionProvider = sessionProvider
-        uiView.goal = goal
-        uiView.setActive(active, animated: true)
-    }
-    
-    func makeCoordinator() -> Coordinator {
-        Coordinator()
-    }
-    
-    class Coordinator: NSObject, ARCoachingOverlayViewDelegate {}
-}
